Add application project

+++ EnvironmentMonitoringSystem.Domain/Models/Event.cs
@@ -5,8 +5,8 @@
         public Guid Id { get; set; }
         public Guid DeviceId { get; set; }
         public DateTime TimeStamp { get; set; }
-        public decimal Temperature { get; set; }
-        public decimal Humidity { get; set; }
+        public double Temperature { get; set; }
+        public double Humidity { get; set; }
         public bool IsAlarm { get; set; }
 
         public Device? Device { get; set; }
